refactoring(Groups) splite service to smallers
