+++ server/src/services/group/index.ts
@@ -1,0 +1,4 @@
+export { GroupFacadeService as GroupService } from './group-facade.service.js';
+export { GroupEnricher } from './group-enricher.service.js';
+export { GroupMembershipService } from './group-membership.service.js';
+export { GroupValidatorService } from './group-validator.service.js';
