+++ server/src/services/index.ts
@@ -1,2 +1,2 @@
 export * from './game.service.js';
-export * from './group.service.js'; 
+export * from './group/index.js';
